Adding utility to print which keyboards are vial-enabled.
This is intended to create an automatically up-to-date list for the
docs. The current solution is just the bash one-liner I thought of
first: it prints the directories that have a /keymap/vial
subdirectory (quite nicely, through the use of tree). This gives a
complete, not efficient, output when a keyboard has variants.